+++ src/modules/CaaSMapper.ts
@@ -431,6 +431,9 @@ export class CaaSMapper {
       ),
       data: await this.mapDataEntries(pageRef.page.formData, [...path, 'data']),
       meta: await this.mapDataEntries(pageRef.page.metaFormData, [...path, 'meta']),
+      ...(pageRef.metaFormData && {
+        metaPageRef: await this.mapDataEntries(pageRef.metaFormData, [...path, 'metaPageRef']),
+      }),
     }
   }
 
@@ -536,6 +539,7 @@ export class CaaSMapper {
       entityType: dataset.entityType,
       data: await this.mapDataEntries(dataset.formData, [...path, 'data']),
       route: dataset.route,
+      routes: dataset.routes,
       template: dataset.template?.uid,
       children: [],
     }